requirements.txt
Brotli==1.0.9
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dash==1.16.0
dash-bootstrap-components==0.10.6
dash-core-components==1.12.0
dash-html-components==1.1.1
dash-renderer==1.8.0
dash-table==4.10.1
firebase-admin==4.5.0
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
google-api-core==1.22.2
google-api-python-client==2.12.0
google-auth==1.21.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.39.0
gunicorn==20.0.4
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
numpy==1.21.1
oauthlib==3.1.1
pandas==1.1.0
plotly==4.9.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
